Moved the CSS Style to an external include
So that users do not have to edit 10+ HTML files to change the colors of their dark theme, I've moved the CSS style into an include file that is now linked into all HTML files in the application. Changes only need to be made in the baseballStyle.css file to use the colors of your dark theme.