Added MySQL/MariaDB Client Connector
Updated the system to use the new C/C++ Client Connector functions to process SQL calls and results. Also added an install process, where now QB64 is compiled as a temporary file to compile the source code if the user can't use the provided binaries. QB64 was added to this repo.